fix: raise technical bar based on intelligence report
