+++ integrations/langgraph/memanto_langgraph/pyproject.toml
@@ -1,10 +1,9 @@
 [project]
 name = "memanto-langgraph"
 version = "0.1.0"
-description = "Type-safe Memanto store integration for LangGraph V3"
+description = "Type-safe Memanto integration for LangGraph V3 BaseStore"
 dependencies = [
-    "langgraph",
-    "langchain-core",
-    "pydantic",
+    "langgraph>=0.1.0",
+    "pydantic>=2.0.0",
     "memanto"
 ]
