+++ integrations/langgraph/memanto_langgraph/schema.py
@@ -1,14 +1,18 @@
-from typing import TypedDict, Annotated, List, Optional, Dict, Any
+from typing import Generic, TypeVar, Any, Dict, Optional
 from pydantic import BaseModel, Field
 
-class MemoryItem(BaseModel):
-    content: str
-    namespace: str
-    metadata: Dict[str, Any] = Field(default_factory=dict)
-    memory_id: Optional[str] = None
+T = TypeVar("T")
 
-class MemantoState(TypedDict):
-    messages: Annotated[List[Any], "The conversation history"]
-    agent_id: str
-    session_id: str
-    context_window: List[MemoryItem]
+class MemantoMemoryItem(BaseModel, Generic[T]):
+    """Type-safe container for Memanto memory entries."""
+    namespace: str
+    key: str
+    value: T
+    metadata: Optional[Dict[str, Any]] = Field(default_factory=dict)
+    version: int = 1
+
+class MemantoStoreConfig(BaseModel):
+    """Configuration for the MemantoStore provider."""
+    api_key: str
+    base_url: str
+    default_namespace: str = "default"
